Inline VisitVolumes
