+++ src/geocel/vg/VecgeomParams.cc
@@ -49,11 +49,11 @@
 #include "corecel/sys/ScopedMem.hh"
 #include "corecel/sys/ScopedProfiling.hh"
 #include "geocel/GeantGeoParams.hh"
+#include "geocel/VolumeVisitor.hh"
 #include "geocel/detail/LengthUnits.hh"
 #include "geocel/detail/MakeLabelVector.hh"
 
 #include "VecgeomData.hh"  // IWYU pragma: associated
-#include "VisitVolumes.hh"
 
 #include "detail/VecgeomCompatibility.hh"
 #include "detail/VecgeomSetup.hh"
@@ -90,6 +90,29 @@ namespace
 #endif
 
 //---------------------------------------------------------------------------//
+class VecgeomVolumeAccessor final
+    : public VolumeAccessorInterface<vecgeom::LogicalVolume const*,
+                                     vecgeom::VPlacedVolume const*>
+{
+  public:
+    //! Outgoing volume node from an instance
+    VolumeRef volume(VolumeInstanceRef parent) final
+    {
+        CELER_EXPECT(parent);
+        auto result = parent->GetLogicalVolume();
+        CELER_ENSURE(result);
+        return result;
+    }
+
+    //! Outgoing instance nodes from a volume
+    ContainerVolInstRef children(VolumeRef parent) final
+    {
+        auto&& daughters = parent->GetDaughters();
+        return ContainerVolInstRef(daughters.begin(), daughters.end());
+    }
+};
+
+//---------------------------------------------------------------------------//
 // HELPER FUNCTIONS
 //---------------------------------------------------------------------------//
 /*!
@@ -118,28 +141,30 @@ std::vector<Label> make_logical_vol_labels(vecgeom::VPlacedVolume const& world)
 {
     using VolT = vecgeom::LogicalVolume;
     std::unordered_map<std::string, std::vector<VolT const*>> names;
-    visit_volumes(
-        [&](VolT const* lv) {
-            CELER_EXPECT(lv);
-            std::string name{lv->GetLabel()};
-            if (starts_with(name, "[TEMP]"))
-            {
-                // Temporary volume not directly used in transport, generated
-                // by g4vg
-                return;
-            }
 
-            // In case it's loaded by vgdml, strip the pointer
-            if (auto pos = name.find("0x"); pos != std::string::npos)
-            {
-                // Copy pointer as 'extension' and delete from name
-                name.erase(name.begin() + pos, name.end());
-            }
+    VolumeVisitor visit_vol{VecgeomVolumeAccessor{}};
+    visit_vol(make_visit_volume_once<vecgeom::LogicalVolume const*>(
+                  [&](VolT const* lv) {
+                      CELER_EXPECT(lv);
+                      std::string name{lv->GetLabel()};
+                      if (starts_with(name, "[TEMP]"))
+                      {
+                          // Temporary volume not directly used in transport,
+                          // generated by g4vg
+                          return;
+                      }
 
-            // Add to name map
-            names[name].push_back(lv);
-        },
-        &world);
+                      // In case it's loaded by vgdml, strip the pointer
+                      if (auto pos = name.find("0x"); pos != std::string::npos)
+                      {
+                          // Copy pointer as 'extension' and delete from name
+                          name.erase(name.begin() + pos, name.end());
+                      }
+
+                      // Add to name map
+                      names[name].push_back(lv);
+                  }),
+              world.GetLogicalVolume());
 
     return detail::make_label_vector<VolT const*>(
         std::move(names), [](VolT const* vol) { return vol->id(); });
@@ -157,7 +182,8 @@ std::vector<Label> make_physical_vol_labels(vecgeom::VPlacedVolume const& world)
 
     // Visit PVs, mapping names to instances, skipping those that have already
     // been visited at a deeper level
-    visit_volume_instances(
+    VolumeVisitor visit_vol{VecgeomVolumeAccessor{}};
+    visit_vol(
         [&](VolT const* pv, int depth) {
             auto&& [iter, inserted] = max_depth.insert({pv, depth});
             if (!inserted)
@@ -382,6 +408,8 @@ VecgeomParams::VecgeomParams(vecgeom::GeoManager const& geo,
         }
     }
     {
+        CELER_LOG(status) << "Constructing metadata";
+
         // Save host data
         HostVal<VecgeomParamsData> host_data;
         host_data.scalars.host_world = geo.GetWorld();
